% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.r
\name{clonalOverlap}
\alias{clonalOverlap}
\title{repertorie analysis functions used for shinyClone 
modified based on the functions from Borcherding N's scRepertoire R package
Examining the clonal overlap between groups or samples}
\usage{
clonalOverlap(
  Combined,
  group.by,
  method = c("overlap", "morisita", "jaccard", "raw"),
  exportTable = FALSE
)
}
\arguments{
\item{Combined}{The \code{CombinedDataSet} object obtained from \code{getCombinedDataSet} function.}

\item{method}{The method to calculate the overlap, either the "overlap" 
coefficient, "morisita", "jaccard" indices, or "raw" for the base numbers.}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\value{
ggplot of the clonotypic overlap between elements of a list
}
\description{
This functions allows for the calculation and visualizations of the 
overlap coefficient, morisita, or jaccard index for clonotypes. 
The overlap coefficient is calculated using the intersection of clonotypes 
divided by the length of the smallest component.
}
\examples{

Combined <- getCombinedDataSet(TCR,Mice.sub2)
clonalOverlap(combined,method = "overlap",exportTable = FALSE)

}
