% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r
\docType{class}
\name{CombinedDataSet-class}
\alias{CombinedDataSet-class}
\title{Class \code{CombinedDataSet}}
\description{
This \code{CombinedDataSet} class holds the combined scRNA-seq and scTCR-seq information for cells, which are relevant for 
	performing clonotype lineage inference and visualization.
}
\section{Slots}{

\describe{
\item{\code{seurat.}}{A seurat object containing the UMAP coordinates, cluster labels and clonotype information for cells
\itemize{
\item{cdr3} {character. Vector of CDR3 sequence that are used to define cell clonotype }
\item{cell.id}{character. Vector of cell barcodes}
\item{clusters}{character or numeric. A vector of character or numeric values indicating clustering labels for cells}
\item{Other terms} {other metadata stored in the Seurat object}}}
}}

