% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineG.R
\name{determineG}
\alias{determineG}
\title{Determine Groups}
\usage{
determineG(n, m)
}
\arguments{
\item{n}{number of observations in dataset}

\item{m}{number of successes in dataset}
}
\value{
returns appropriate group size or an error upon failure
}
\description{
\code{determineG} Determines the appropriate number of groups when performing the Hosmer-Lemeshow test.
}
\details{
This function takes a sample size (n) and number of success (m) to determine the appropriat number of groups to use with the Hosmer-Lemeshow test.
}
\examples{
determineG(n=10000,m=500)


}

