% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{expression2List_seurat}
\alias{expression2List_seurat}
\title{Allows users to take the meta data in Seurat and place it into a list 
that will work with all the functions}
\usage{
expression2List_seurat(seurat, group.by)
}
\arguments{
\item{seurat}{a seurat object.}

\item{group.by}{The column header to group the new list. NULL will return clusters.}
}
\value{
list derived from the meta data of seurat object with 
elements divided by the group.by parameter
}
\description{
Allows users to perform more fundamental measures of clonotype analysis 
using the meta data from the Seurat object. For Seurat objects the 
active identity is automatically added as "cluster". Remaining grouping 
parameters or Seurat objects must appear in the meta data.
}
\examples{
Combined <- getCombinedDataSet(TCR, Mice.sub2)
newList <- expression2List(Combined, group.by = "seurat_clusters")

}
