% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.r
\name{getCirclize}
\alias{getCirclize}
\title{a function to prepare input for chord diagram}
\usage{
getCirclize(Combined, group.by, proportion)
}
\arguments{
\item{group.by}{The group header for which you would like to analyze 
the data.}

\item{proportion}{Binary will calculate relationship unique 
clonotypes (proportion = FALSE) or a ratio of the group.by 
variable (proportion = TRUE)}

\item{sc}{object after combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}
}
\value{
data frame of shared clonotypes between groups
}
\description{
a function to prepare input for chord diagram
}
\author{
Dillon Corvino, Nick Borcherding
}
