% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r
\name{getClonotypeLineages}
\alias{getClonotypeLineages}
\title{Infer clonotype lineage structure from \code{CombinedDatSet}}
\usage{
getOverallTrajectory(CombinedDataSet, ...)
}
\arguments{
\item{CombinedDataSet}{The \code{CombinedDataSet} object obtained from \code{getCombinedDataSet}}

\item{...}{Additional arguments to specify how lineages are constructed from clusters}
}
\description{
Infer clonotype lineage structure from \code{CombinedDatSet}
}
