% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r
\docType{methods}
\name{getCombinedDataSet}
\alias{getCombinedDataSet}
\title{Initialize an object of class \code{CombinedDataSet}}
\usage{
getCombinedDataSet(TCR, seurat, ...)
}
\arguments{
\item{seurat}{An integrated Seurat object. The object must contain clustering results and saved as 'clusters', and must have cell barcodes saved as 'cell.id'}

\item{scTCRseq}{The preprocessed scTCR-seq data. See details}
}
\description{
Construct a \code{CombinedDataSet} object with Seurat clustering results and clonotype information . Additional helper methods 
 for manipulating \code{CombinedDatSet} objects are also described below.
}
\details{
The input scTCR-seq needs to be a dataframe, and alpha chain and beta chain need to be paired. If you have multiple sets of scTCR-seq from more than one sample, the data needs to be combined into one.

The seurat object must contain clustering results and saved as 'clusters', contain cell barcodes saved as 'cell.id' and also need to have UMAP coordinates for cells.
}
