% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverallTrajectory.R
\name{getOverallTrajectory,CombinedDataSet-method}
\alias{getOverallTrajectory,CombinedDataSet-method}
\title{get Overall Trajectory}
\usage{
\S4method{getOverallTrajectory}{CombinedDataSet}(
  CombinedDataSet,
  reducedDim = "UMAP",
  start.clus = NULL,
  end.clus = NULL,
  dist.method = "slingshot",
  use.median = FALSE,
  ...
)
}
\arguments{
\item{reducedDim}{The dimensionality reduction to be used. Can be
a matrix or a character identifying which element of
\code{reducedDim(data)} is to be used. If multiple dimensionality
reductions are present and this argument is not provided, the first element
will be used by default.}

\item{start.clus}{(optional) character, indicates the starting cluster(s)
from which lineages will be drawn.}

\item{end.clus}{(optional) character, indicates which cluster(s) will be
forced to be leaf nodes in the graph.}

\item{dist.method}{(optional) character, specifies the method for calculating
distances between clusters. Default is \code{"slingshot"}, see
\code{\link[TrajectoryUtils]{createClusterMST}} for details.}

\item{use.median}{logical, whether to use the median (instead of mean) when
calculating cluster centroid coordinates. Default is \code{'TRUE'}.}

\item{data}{A class of \code{CombinedDataSet} object}
}
\value{
A class of \code{TrajectoryDataSet} object
}
\description{
This function infers the overall trajectory structure from \code{CombinedDataSet} using \code{\link[slingshot]}.
}
\details{
Given a \code{CombinedDataSet} object, this function infers the overall cell trajectory using the \code{\link[slingshot]}.
  To reduce noise, clones with less than 5 cells are removed before trajectory inference.
}
\examples{
TCR <-read.csv('/PATH/TO/YOUR/scTCR-seqData/',header=T)
load('Mice.sub.rda')
Combined <- getCombinedDataSet(TCR,Mice.sub)
Trajectory <- getOverallTrajectory(Combined,reducedDim = NULL,start.clus = NULL, end.clus = NULL, dist.method = 'slingshot', use.median = TRUE)

}
