% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r
\name{getShinyInput}
\alias{getShinyInput}
\title{prepare the input for shiny apps}
\usage{
getShinyInput(CombinedDataSet, ...)
}
\arguments{
\item{CombinedDataSet}{The \code{CombinedDataSet} object}
}
\description{
prepare the input for shiny apps
}
