% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r, R/HelperFuntions.r
\name{getTCRs}
\alias{getTCRs}
\alias{getTCRs,CombinedDataSet-method}
\title{extract the TCR information fomr CombinedDataSet}
\usage{
getTCRs(CombinedDataSet, ...)

\S4method{getTCRs}{CombinedDataSet}(CombinedDataSet)
}
\description{
extract the TCR information fomr CombinedDataSet
}
