% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{grabMeta}
\alias{grabMeta}
\title{This is to grab the meta data from a CombinedDataSet object}
\usage{
grabMeta(Combined, group.by)
}
\arguments{
\item{Combined}{a \code{CombinedDataSet}object.}

\item{group.by}{The column header to group the new list. NULL will return clusters.}
}
\description{
This is to grab the meta data from a CombinedDataSet object
}
