% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{grabMeta_seurat}
\alias{grabMeta_seurat}
\title{This is to grab the meta data from a seurat object}
\usage{
grabMeta_seurat(seurat, group.by)
}
\arguments{
\item{group.by}{The column header to group the new list. NULL will return clusters.}

\item{DATA}{a Seurat object.}
}
\description{
This is to grab the meta data from a seurat object
}
