% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{jaccardIndex}
\alias{jaccardIndex}
\title{Calculate the Jaccard Similarity Index for Overlap Analysis}
\usage{
jaccardIndex(df, length, coef_matrix)
}
\description{
Calculate the Jaccard Similarity Index for Overlap Analysis
}
