% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r, R/HelperFuntions.r
\name{lrtLineages}
\alias{lrtLineages}
\alias{lrtLineages,TrajectoryDataSet-method}
\title{extract the inferred clonotype lineages}
\usage{
lrtLineages(TrajectoryDataSet, ...)

\S4method{lrtLineages}{TrajectoryDataSet}(TrajectoryDataSet)
}
\description{
extract the inferred clonotype lineages
}
