% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r, R/HelperFuntions.r
\name{lrtParams}
\alias{lrtParams}
\alias{lrtParams,TrajectoryDataSet-method}
\title{extract the parameters used when estimating lineages}
\usage{
lrtParams(TrajectoryDataSet, ...)

\S4method{lrtParams}{TrajectoryDataSet}(TrajectoryDataSet)
}
\description{
extract the parameters used when estimating lineages
}
