% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{morisitaIndex}
\alias{morisitaIndex}
\title{calculate the Morisita index for Overlap Analysis}
\usage{
morisitaIndex(df, length, coef_matrix)
}
\description{
calculate the Morisita index for Overlap Analysis
}
