% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.r
\name{occupiedscRepertoire}
\alias{occupiedscRepertoire}
\title{Visualize the number of single cells with clonotype frequencies by cluster}
\usage{
occupiedscRepertoire(
  Combined,
  group.by = NULL,
  cloneTypes = c(Small = 0.001, Medium = 0.01, Large = 0.1, Hyperexpanded = 1),
  proportion = TRUE,
  label = TRUE,
  facet.by = NULL,
  na.include = FALSE,
  exportTable = FALSE
)
}
\arguments{
\item{proportion}{Convert the stacked bars into relative proportion}

\item{label}{Include the number of clonotype in each category by x.axis variable}

\item{facet.by}{The column header used for faceting the graph}

\item{na.include}{Visualize NA values or not.}

\item{exportTable}{Exports a table of the data into the global 
environment in addition to the visualization}

\item{sc}{The seurat or SCE object to visualize after combineExpression(). 
For SCE objects, the cluster variable must be in the meta data under 
"cluster".}

\item{x.axis}{The variable in the meta data to graph along the x.axis}
}
\value{
Stacked bar plot of counts of cells by clonotype frequency group
}
\description{
View the count of clonotypes frequency group in seurat or SCE object 
meta data after combineExpression(). The visualization will take the 
new meta data variable "cloneType" and plot the number of cells with
each designation using a secondary variable, like cluster. Credit to 
the idea goes to Drs. Carmona and Andreatta and their work with
\href{https://github.com/carmonalab/ProjecTILs}{ProjectTIL}.
}
\examples{
Combined <- getCombinedDataSet(TCR, Mice.sub2)
occupiedscRepertoire(Combined,group.by = 'clusters', cloneTypes=c(Small = 0.001, Medium = 0.01, Large = 0.1, Hyperexpanded = 1),proportion = TRUE,
 label = TRUE,facet.by = NULL, na.include = FALSE,exportTable = FALSE)
}
