% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{overlapIndex}
\alias{overlapIndex}
\title{Calculate the Overlap Coefficient for Overlap Analysis}
\usage{
overlapIndex(df, length, coef_matrix)
}
\description{
Calculate the Overlap Coefficient for Overlap Analysis
}
\author{
Nick Bormann, Nick Borcherding
}
