% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r, R/HelperFuntions.r
\name{seuratClusterLabels}
\alias{seuratClusterLabels}
\alias{seuratClusterLabels,CombinedDataSet-method}
\title{extract the seurat cell cluster labels from CombinedDataSet}
\usage{
seuratClusterLabels(CombinedDataSet, ...)

\S4method{seuratClusterLabels}{CombinedDataSet}(CombinedDataSet)
}
\arguments{
\item{CombinedDataSet}{}
}
\description{
extract the seurat cell cluster labels from CombinedDataSet
}
