% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyClone.r
\name{shinyClone}
\alias{shinyClone}
\title{Explanatory analysis of clonotype}
\usage{
shinyClone(CombinedDataSet = NULL)
}
\arguments{
\item{Combined}{a \code{CombinedDataSet} object}
}
\value{
shiny app
}
\description{
Explanatory analysis of clonotype
}
