% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r
\docType{class}
\name{TrajectoryDataSet-class}
\alias{TrajectoryDataSet-class}
\title{Class \code{TrajectoryDataSet}}
\description{
This \code{TrajectoryDataSet} class holds the inferred clonotype lineages from \code{getOverallTrajectory}.
}
\section{Slots}{

\describe{
\item{\code{clonotype.}}{A vector of clonotype that has lineages built}

\item{\code{lineages.}}{A list of lineages corresponding the the clonotypes}

\item{\code{trajectoryParams.}}{Additional parameters used by \code{getOverallTrajectory}. These may specify how the minimum spanning tree on clusters was constructed:
\itemize{
\item{\code{start.clus}}{ character. The label of the root cluster, or a vector of cluster labels giving the root clusters of each disjoint component of th egraph}
\item{\code{end.clust}}{ character. Vector of cluster labels indicating terminal clusters.}
\item{\code{dist.method}}{character. Specify the method for calculating distances between clusters}
\item{\code{use.median}}{logical. Whether to use the median instead of mean when calculating cluster centroid coordinates. Default is \code{'TRUE'}.}
\item{Other parameters specified by \code{\link[princurve]{principal_curve}}}.}}
}}

\examples{
# TrajectoryDataSet objects are generated from \code{getOverallTrajectory}
}
